requirements.txt
streamlit==1.36.0
google-generativeai==0.3.1 # Or a compatible version you prefer
pdfplumber==2.0.4          # As discussed, this is a stable version
pandas==2.2.0              # Or a compatible version
openpyxl==3.1.2            # Necessary for pandas to write .xlsx files